[Refactor] Move OS system metrics into Platform
