+++ be/src/io/core/io_profiler.cpp
@@ -14,6 +14,7 @@
 
 #include "io/core/io_profiler.h"
 
+#include <algorithm>
 #include <cstdint>
 #include <memory>
 #include <mutex>
