+++ be/src/service/backend_metrics_initializer.cpp
@@ -41,6 +41,7 @@
 #include "fs/key_cache.h"
 #include "http/http_metrics.h"
 #include "io/core/io_profiler_metrics.h"
+#include "platform/platform_metrics.h"
 #include "runtime/runtime_metrics.h"
 #include "runtime/stream_load/stream_load_metrics.h"
 #include "service/service_metrics.h"
@@ -266,7 +267,8 @@ void BackendMetricsInitializer::initialize(ProcessMetricsRegistry* process_metri
     agent_metrics->install_disk_path_metrics(registry, options.storage_paths);
 
     if (options.init_system_metrics) {
-        SystemMetrics::instance()->install(registry, disk_devices, network_interfaces);
+        PlatformMetrics::instance()->install(registry, disk_devices, network_interfaces);
+        SystemMetrics::instance()->install(registry);
         IOProfilerMetrics::instance()->install(registry);
     }
