+++ be/src/util/system_metrics.h
@@ -25,15 +25,9 @@
 
 namespace starrocks {
 
-class CpuMetrics;
-class DiskMetrics;
-class NetMetrics;
-class FileDescriptorMetrics;
-class SnmpMetrics;
 class QueryCacheMetrics;
 class VectorIndexCacheMetrics;
 class RuntimeFilterMetrics;
-class VectorIndexCacheMetrics;
 
 class MemoryMetrics {
 public:
@@ -83,45 +77,18 @@ public:
 
     static SystemMetrics* instance();
 
-    // install system metrics to registry
-    void install(MetricRegistry* registry, const std::set<std::string>& disk_devices,
-                 const std::vector<std::string>& network_interfaces);
+    // install higher-level system metrics to registry
+    void install(MetricRegistry* registry);
 
     // update metrics
     void update();
 
-    void get_disks_io_time(std::map<std::string, int64_t>* map);
-    int64_t get_max_io_util(const std::map<std::string, int64_t>& lst_value, int64_t interval_sec);
-
-    void get_network_traffic(std::map<std::string, int64_t>* send_map, std::map<std::string, int64_t>* rcv_map);
-    void get_max_net_traffic(const std::map<std::string, int64_t>& lst_send_map,
-                             const std::map<std::string, int64_t>& lst_rcv_map, int64_t interval_sec,
-                             int64_t* send_rate, int64_t* rcv_rate);
     const MemoryMetrics* memory_metrics() const { return _memory_metrics.get(); }
 
     void update_memory_metrics();
 
 private:
-    void _install_cpu_metrics(MetricRegistry*);
-    // On Intel(R) Xeon(R) CPU E5-2450 0 @ 2.10GHz;
-    // read /proc/stat would cost about 170us
-    void _update_cpu_metrics();
-
     void _install_memory_metrics(MetricRegistry* registry);
-
-    void _install_disk_metrics(MetricRegistry* registry, const std::set<std::string>& devices);
-    void _update_disk_metrics();
-
-    void _install_net_metrics(MetricRegistry* registry, const std::vector<std::string>& interfaces);
-    void _update_net_metrics();
-
-    void _install_fd_metrics(MetricRegistry* registry);
-
-    void _update_fd_metrics();
-
-    void _install_snmp_metrics(MetricRegistry* registry);
-
-    void _update_snmp_metrics();
 
     void _install_query_cache_metrics(MetricRegistry* registry);
 
@@ -141,20 +108,12 @@ private:
 private:
     static const char* const _s_hook_name;
 
-    std::unique_ptr<CpuMetrics> _cpu_metrics;
     std::unique_ptr<MemoryMetrics> _memory_metrics;
-    std::map<std::string, DiskMetrics*> _disk_metrics;
-    std::map<std::string, NetMetrics*> _net_metrics;
-    std::unique_ptr<FileDescriptorMetrics> _fd_metrics;
     std::unique_ptr<QueryCacheMetrics> _query_cache_metrics;
     std::unique_ptr<VectorIndexCacheMetrics> _vector_index_cache_metrics;
     std::map<std::string, RuntimeFilterMetrics*> _runtime_filter_metrics;
-    int _proc_net_dev_version = 0;
-    std::unique_ptr<SnmpMetrics> _snmp_metrics;
 
     std::mutex _update_mutex;
-    char* _line_ptr = nullptr;
-    size_t _line_buf_size = 0;
     MetricRegistry* _registry = nullptr;
 };
